Decrease 5ms delay in atapi transfers to 50us for Sam460ex drivers.
The large delay is responsible for slow loading times with the standard
slb of AmigaOS. It seems that the slb issues a command for each block, thus
each transfer could last at up to 10ms.

By decreasing the delay to 50us (which is also used in some other places),
the speed of loading kickstart is now about 100 times faster.

According to the comments, the delay makes sure that the hardware sets the
busy signal, on which the code waits to be cleared. However, for QEMU the
delays seem to be unnecessary altogether, so they could be removed. It is
unclear how this change affects real hardware, so we keep at least the 50us
delays in case anyone want to try this with real hardware.

Similar delays in other places are not changes by this patch.